+++ spec/common/subline_object_spec.rb
@@ -24,7 +24,7 @@ describe 'A Subline Object' do
   end
 
   after :all do
-    @ole.quit
+    @ole.quit unless @ole.nil?
   end
 
   it 'should create a new instance' do